Separate logic into helper NextSessionUtil class
